Hackaround until we release this bundle
Right now this bundle subclasses the DROLSKY bundle and gets
Code::TidyAll::Test::Vars added a develop prereq by doing so.

However, this bundle is built using itself. On Travis, that means it's built
using the _prior_ version of the MAXMIND bundle, which does not require CTVA,
so we need to require it explicitly until after the release.